Moving mappers to separate package
