+++ backend/db/src/main/java/com/bls/patronage/db/mappers/DeckWithFlashcardsNumberMapper.java
@@ -1,4 +1,4 @@
-package com.bls.patronage.db.dao;
+package com.bls.patronage.db.mappers;
 
 import com.bls.patronage.db.model.DeckWithFlashcardsNumber;
 import org.skife.jdbi.v2.StatementContext;
